rag/pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel>=0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gryphgen-rag"
version = "1.0.0"
description = "Modern implementation of SimGRAG and CAG: Advanced Retrieval-Augmented Generation methods"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "GRYPHGEN Contributors", email = "info@gryphgen.ai"}
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "knowledge-graphs",
    "llm",
    "ai",
    "machine-learning",
    "simgrag",
    "cag",
    "cache-augmented-generation",
    "nvidia",
    "gpu"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "torch>=2.5.0",
    "transformers>=4.47.0",
    "sentence-transformers>=3.3.0",
    "networkx>=3.4.2",
    "ollama>=0.4.4",
    "sqlalchemy>=2.0.36",
    "numpy>=2.1.3",
    "pydantic>=2.10.3",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
    "rich>=13.9.4",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.48",
    "mkdocstrings>=0.27.0",
]
gpu = [
    "faiss-gpu>=1.9.0",
    "cupy-cuda12x>=13.3.0",
    "triton>=3.2.0",
]
all = [
    "gryphgen-rag[dev,docs,gpu]",
]

[project.urls]
Homepage = "https://github.com/danindiana/GRYPHGEN"
Documentation = "https://github.com/danindiana/GRYPHGEN/tree/main/rag"
Repository = "https://github.com/danindiana/GRYPHGEN"
Issues = "https://github.com/danindiana/GRYPHGEN/issues"

[project.scripts]
gryphgen-rag = "gryphgen_rag.cli:main"

[tool.setuptools]
packages = ["simgrag", "cag", "common", "gryphgen_rag"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "networkx.*",
    "ollama.*",
    "transformers.*",
    "torch.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=simgrag",
    "--cov=cag",
    "--cov=common",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["simgrag", "cag", "common"]
omit = [
    "*/tests/*",
    "*/examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
